Remove nested event array.
Removes a race condition where calling the `Scheduler.clear` followed by
`Scheduler.schedule` from within a called to `Scheduler.check` would
case `check` to send an `unset` event before exit unsetting the timer in
`Timer` so that the schedule event would not fire. This was a race
condition because it would only occur when the `check` was called the
exact millisecond that the event was scheduled and the event scheduled
from within `check` was scheduled for the current millisecond.

The problem was that we where losing trace of the nested array of
events. Working through an array of events for the current millisecond,
but pushing events for the current millisecond onto an array held by a
newly vivified entry in the event lookup tree. We now put the event in
the tree using a compound key of the event time and the event key
(name) so that the check loop continually checks the min value in the
event tree instead of working through an array. All the functions
reference the event tree, no array to get lost to improper indirection.